requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
gradio>=4.0.0
Pillow>=10.0.0
nicegui>=2.0.0
